Refactor revision perform and log logic
database.Run has been renamed to database.Perform to keep the API
consistent with the logic used in both cmd.Run and cmd.Reset.

The revision hash is now going to be stored as a binary object in the
database to keep things simple as it's used as a []byte in the program
itself.