Add support for multiple databases
This is achieved by the addition of the -c, --config flag which will
change the config directory that is used for managing a databases
revisions. The idea being that if you have multiple databases you wish
to run migrations against you simple call the init command multiples
times.

  mgrt init prod
  mgrt init uat

Like so, then change the directory to interact with via the -c flag.

  mgrt add -m "Create users table" -c uat
  mgrt run -c uat