Remove authorship from revisions and use TOML over YAML
Having authorship in the maing configuration file doesn't make much
sense. If I am going to go back to add authorship to revisions, which
down the line might make sense, this it should be pulled from the
environment. The question is where, have a ~/.mgrt.toml file containing
the author details, or perhaps pull it from Git, if it is being used
with git.